fix(mcp): preserve stream selectors and classify compatibility coverage
